+++ packages/fs-syncer/src/__tests__/fixtures/integration.test.ts/adds-newline-by-default/test.txt
@@ -1,0 +1,1 @@
+abc
