+++ packages/fs-syncer/src/index.ts
@@ -1,105 +1,14 @@
 import * as fs from 'fs'
 import * as path from 'path'
-import * as os from 'os'
-import {getPaths, get, dedent} from './util'
+import {getPaths, get} from './util'
+import {fsSyncerFileTreeMarker, CreateSyncerParams} from './types'
+import {defaultMergeStrategy, defaultBeforeWrites} from './defaults'
 
-// todo: make .read() have a more helpful return type
-// todo: mergeStrategy. tried before
-// todo: fixture
+export * from './types'
+export * from './defaults'
+export {jestFixture} from './jest'
 
-const fsSyncerFileTreeMarker = Symbol('fs-syncer-marker')
-
-export type FsSyncerFile =
-  | string
-  | {
-      [fsSyncerFileTreeMarker]: () => string
-    }
-
-export type MergeStrategy = (params: {
-  filepath: string
-  existingContent: string | undefined
-  targetContent: string | undefined
-}) => string | undefined
-
-export type BeforeWrite = (params: {filepath: string; content: string}) => string
-
-export const defaultMergeStrategy: MergeStrategy = params => params.targetContent
-
-export const defaultBeforeWrites: BeforeWrite[] = [
-  params => dedent(params.content),
-  params => params.content.trim() + os.EOL,
-]
-
-export interface CreateSyncerParams<T extends object> {
-  /** Path that all files will be synced relative to. */
-  baseDir: string
-  /**
-   * Object representation of desired file tree.
-   *
-   * @example
-   * ```
-   * {
-   *   'one.txt': 'uno',
-   *   'two.txt': 'dos',
-   *   'subfolder': {
-   *     'three.txt': 'tres'
-   *   }
-   * }
-   * ```
-   */
-  targetState: T
-  /**
-   * Any file or folder matching one of these regexes will not be read from the file system
-   *
-   * @default
-   * ```
-   * ['node_modules']
-   * ```
-   *
-   * @example
-   * ```
-   * ['node_modules', 'dist']
-   * ```
-   *
-   * @example
-   * ```
-   * ['node_modules', 'dist', /.*\.(test|spec)\.ts/]
-   * ```
-   *
-   * @example
-   * ```
-   * /^((?!src).)*$/ // ignore all paths not containing `src`. Note - this effectively means paths must _start_ with `src`, since they will be short-circuited otherwise.
-   * ```
-   */
-  exclude?: Array<string | RegExp>
-
-  /**
-   * A function which takes a filepath, old content and new content strings, and returns a string. The returned string is written to disk.
-   * If `undefined` is returned, the file is deleted.
-   *
-   * You can use this to implement custom merges. e.g. a json config file with some default properties could use a strategy which merges existing and target contents:
-   *
-   * @example
-   * ```
-   * ({filepath, targetContent, existingContent}) => {
-   *   if (!existingContent) {
-   *     return targetContent
-   *   }
-   *   const existingConfig = JSON.parse(existingContent)
-   *   const targetConfig = JSON.parse(targetContent || '{}')
-   *
-   *   return {
-   *     ...targetConfig,
-   *     ...existingConfig,
-   *   }
-   * }
-   * ```
-   *
-   * @default params => params.targetContent
-   */
-  mergeStrategy?: MergeStrategy
-  beforeWrites?: BeforeWrite[]
-}
+export const isFsSyncerFileTree = (obj: any): boolean => Boolean(obj?.[fsSyncerFileTreeMarker])
 
 const tryCatch = <T, U = undefined>(fn: () => T, onError: (error: unknown) => U = () => (undefined as any) as U) => {
   try {
@@ -155,7 +64,7 @@ export const createFSSyncer = <T extends object>({
     return result
   }
 
-  const read = (): any => fs.existsSync(baseDir) ? readdir(baseDir) : {}
+  const read = (): any => (fs.existsSync(baseDir) ? readdir(baseDir) : {})
 
   /** writes all target files to file system, and deletes files not in the target state object */
   const sync = () => {
@@ -185,81 +94,6 @@ export const createFSSyncer = <T extends object>({
 
   return syncer
 }
-
-/**
- * @experimental
- * Call from a jest test to setup a syncer in a `baseDir` based on the current file, suite and test name.
- * This reduces the risk of copy-paste errors resulting in two tests trying to write to the same directory.
- * @param targetState target file tree
- */
-export const jestFixture = Object.assign(
-  (targetState: object) => {
-    return createFSSyncer<any>({
-      baseDir: jestFixture.baseDir(),
-      targetState,
-    })
-  },
-  {
-    baseDir: () =>
-      path.join(
-        path.dirname(expect.getState().testPath),
-        'fixtures',
-        path.basename(expect.getState().testPath),
-        expect
-          .getState()
-          .currentTestName.toLowerCase()
-          .replace(/[^\da-z]/g, '-')
-          .replace(/-+/g, '-')
-          .replace(/^-*/, '')
-          .replace(/-*$/, '')
-      ),
-    addYamlSnapshotSerializer: (indent = '  ') => {
-      expect.addSnapshotSerializer({
-        test: isFsSyncerFileTree,
-        print: val => yamlishPrinter(val, indent),
-      })
-    },
-    wipe: () =>
-      createFSSyncer({
-        baseDir: path.join(path.dirname(expect.getState().testPath), 'fixtures'),
-        targetState: {},
-      }).sync(),
-  }
-)
-
-const yamlishPrinter = (val: any, tab = '  ') => {
-  const buffer: string[] = []
-  const printNode = (node: any, indent: number) => {
-    if (typeof node === 'undefined') {
-      return
-    }
-    if (node && typeof node === 'object') {
-      const entries = Object.entries(node).sort((...items) => {
-        const keys = items.map(e => (e[1] && typeof e[1] === 'object' ? 'z' : typeof e[1]))
-        return keys[0].localeCompare(keys[1])
-      })
-      entries.forEach(e => {
-        buffer.push('\n' + tab.repeat(indent) + e[0] + ': ')
-        printNode(e[1], indent + 1)
-      })
-      return
-    }
-    if (typeof node === 'string' && node.includes('\n')) {
-      buffer.push('|-\n')
-      node.split('\n').forEach((line, i, arr) => {
-        buffer.push(tab.repeat(indent) + line + (i === arr.length - 1 ? '' : '\n'))
-      })
-      return
-    }
-
-    buffer.push(node?.toString())
-  }
-
-  printNode(val, 0)
-  return buffer.join('').trimLeft()
-}
-
-const isFsSyncerFileTree = (obj: any): boolean => Boolean(obj?.[fsSyncerFileTreeMarker])
 
 /**
  * A helper to read and write text files to a specified directory.
